split collector as a multi-module crate
