+++ collector/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dd-trace-collector"
+name = "ddcollector"
 version = "0.1.0"
 edition = "2021"
 
@@ -13,6 +13,8 @@ strip = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.64"
+hashbrown = "0.13.2"
 hyper = { version = "0.14.24", features = ["full"] }
 rmp = "0.8.11"
 tokio = { version = "1.25.0", features = ["full"] }
