+++ collector/src/exporting.rs
@@ -1,0 +1,9 @@
+use crate::tracing::{Traces};
+use async_trait::async_trait;
+
+pub mod agent;
+
+#[async_trait]
+pub trait Exporter {
+    async fn export(&self, traces: Traces);
+}
